// frontend/src/components/FearGreedIndex.d.ts
import React from 'react';
interface Props {
    data?: {
        composite_score?: number;
        value?: number;
        classification?: string;
        timestamp?: string;
    };
    loading?: boolean;
}
declare const FearGreedIndex: React.FC<Props>;
export default FearGreedIndex;
//# sourceMappingURL=FearGreedIndex.d.ts.map